update: foi refatorado novamente o app.tsx para uma melhor estruturação e modificação do arquivo layout e criação de uma unica interface de componente
