+++ vite-project/src/components/layout/index.tsx
@@ -1,0 +1,9 @@
+import SideMenu from "../menu-lateral/menu-lateral";
+
+const Layout = ({ children }) => (
+  <div className="flex">
+    <SideMenu />
+    <div className="flex-1">{children}</div>
+  </div>
+);
+export default Layout;
