--- vite-project/src/utils/interfaces/ProfileCardComponents.tsx
@@ -1,10 +1,0 @@
-export interface ProfileCardProps {
-    name: string;
-    especifico: string;
-    dataNacimento: string;
-    cidade: string;
-    estado: string;
-    profile: string;
-    profissao: string;
-  }
-
